pyproject.toml
[tool.poetry]
name = "repl-nix-mcproject-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_mcproject_1"}]

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
google-api-python-client = "^2.137.0"
google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
